+++ components/admin/pages/form/PagesForm.js
@@ -27,6 +27,7 @@ class PagesForm extends React.Component {
     id: this.props.id,
     loading: !!this.props.id,
     form: STATE_DEFAULT.form,
+    initialForm: STATE_DEFAULT.form,
   });
 
   componentDidMount() {
@@ -41,6 +42,7 @@ class PagesForm extends React.Component {
         .then((data) => {
           this.setState({
             form: this.setFormFromParams(data),
+            initialForm: this.setFormFromParams(data),
             // Stop the loading
             loading: false,
           });
@@ -83,6 +85,7 @@ class PagesForm extends React.Component {
             updatePage(form, authorization)
               .then((data) => {
                 toastr.success('Success', `The page "${data.id}" - "${data.title}" has been updated correctly`);
+                this.setState({ initialForm: form });
                 if (onSubmit) onSubmit();
               })
               .catch((err) => {
@@ -94,6 +97,7 @@ class PagesForm extends React.Component {
             createPage(form, authorization)
               .then((data) => {
                 toastr.success('Success', `The page ${data.title}" has been created correctly`);
+                this.setState({ initialForm: form });
                 if (onSubmit) onSubmit();
               })
               .catch((err) => {
@@ -145,12 +149,14 @@ class PagesForm extends React.Component {
   }
 
   render() {
-    const { form, loading, id, step, stepLength, submitting } = this.state;
+    const {
+      form, loading, id, step, stepLength, submitting, initialForm,
+    } = this.state;
     return (
       <form
         className={cx({
           'c-form': true,
-          '-disabled': form.env && process.env.NEXT_PUBLIC_ENVS_EDIT.split(',').findIndex((d) => d === form.env) < 0,
+          '-disabled': initialForm.env && process.env.NEXT_PUBLIC_ENVS_EDIT.split(',').findIndex((d) => d === initialForm.env) < 0,
         })}
         onSubmit={this.onSubmit}
         noValidate