+++ pages/dashboards/ocean-watch/partners/index.jsx
@@ -152,13 +152,6 @@ export default function OceanWatchPartnersPage() {
 }
 
 export async function getStaticProps() {
-  // feature flag to avoid display any Ocean Watch development in other environments
-  if (process.env.NEXT_PUBLIC_FEATURE_FLAG_OCEAN_WATCH !== 'true') {
-    return {
-      notFound: true,
-    };
-  }
-
   return {
     props: {},
   };